refactor: 패키지 구조 변경, 변수 이름 변경
