+++ src/main/java/nextstep/courses/domain/session/ImageType.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.session;
 
 public enum ImageType {
     GIF, JPG, JPEG, PNG, SVG
