+++ src/test/java/nextstep/courses/domain/session/sessionType/FreeSessionTypeTest.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.session.sessionType;
 
 import nextstep.payments.domain.Payment;
 import org.junit.jupiter.api.DisplayName;
@@ -10,7 +10,7 @@ class FreeSessionTypeTest {
     FreeSessionType freeSessionType = new FreeSessionType();
 
     @Test
-    @DisplayName("isSessionFull(): currentNumberOfEnrollment에 상관 없이 항상 true를 반환한다.")
+    @DisplayName("isSessionFull(): numberOfCurrentEnrollment에 상관 없이 항상 true를 반환한다.")
     void testIsSessionFull() {
         assertThat(freeSessionType.isSessionFull(10000)).isTrue();
     }
